+++ app/actors/tax/decl.go
@@ -12,8 +12,6 @@ const (
 	ConstErrorLevel  = env.ConstErrorLevelActor
 
 	ConstPriorityValue = 2.50
-
-	ConstProductTaxableAttribute = "taxable"
 )
 
 var priority float64
