+++ app/actors/tax/taxable/decl.go
@@ -1,0 +1,10 @@
+package taxable
+
+const (
+	ConstErrorModule = "taxable"
+	ConstProductTaxableAttribute = "taxable"
+)
+
+// DefaultTaxable is a default implementer of InterfaceTax
+type DefaultTaxable struct{}
+
